requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.7
python-decouple==3.8
django-cors-headers==4.3.1
drf-yasg==1.21.7
Pillow==10.0.1
